# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenRouter-like AI Gateway API
  description: API specification for the AI Gateway Platform, compatible with OpenAI API.
  version: 0.1.0
servers:
  - url: https://api.your-platform.com/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Local development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []

paths:
  /chat/completions:
    post:
      summary: Create a chat completion
      operationId: createChatCompletion
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatCompletionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: Streamed chunks (SSE)
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '402':
          description: Payment Required (Insufficient credits)
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error

  /models:
    get:
      summary: List available models
      operationId: listModels
      tags:
        - Models
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: list
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Model'

components:
  schemas:
    CreateChatCompletionRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
          description: ID of the model to use (e.g., "llama-3-70b", "gpt-4o")
          example: llama-3-70b
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionMessageParam'
        stream:
          type: boolean
          default: false
          description: If set, partial message deltas will be sent.
        temperature:
          type: number
          default: 1.0
        max_tokens:
          type: integer
        top_p:
          type: number
          default: 1.0
        # Routing Extensions
        routing_strategy:
          type: string
          enum: [lowest_cost, lowest_latency, balanced]
          description: Optional routing strategy override.
        provider_allowlist:
          type: array
          items:
            type: string
          description: Optional list of allowed providers (e.g., ["groq", "deepinfra"]).

    ChatCompletionMessageParam:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string

    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: chat.completion
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                $ref: '#/components/schemas/ChatCompletionMessageParam'
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
            # Billing Extension
            total_cost:
              type: number
              description: Total cost for this request in USD.

    Model:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: model
        created:
          type: integer
        owned_by:
          type: string
        pricing:
          type: object
          properties:
            prompt:
              type: string
              description: Cost per 1M tokens
            completion:
              type: string
              description: Cost per 1M tokens